[Xamarin.Android.Build.Tasks] fix for abi delimiters
Context: https://github.com/xamarin/xamarin-android/commit/901dba8cf03e65789f160927e8576c77924b13ff
Context: https://github.com/xamarin/xamarin-android/commit/decfbccf3e99449c9a8d6afef113c9339f1bc85b

In decfbcc, we broke support for `,` and `%3b` as delimiters for
`$(AndroidSupportedAbis)`. We did not have a test checking for these
cases.

I fixed this in 901dba8c, but the diff was too large to be deemed
"safe" for 16.2 at this point in time.

A one-line patch to fix the issue is this MSBuild incantation:

    <_XARuntimeArchitecture Include="$([MSBuild]::Unescape($(_BuildTargetAbis.Replace(',', ';'))))"/>

We replace the `,` with `;` -- which gets escaped to `%3b`. The
`[MSBuild]::Unescape` call converts `%3b` back to `;`.

This also supports `%3b`.