[tests] Exclude additional tests.
Exclude two categories of tests:

 1. `Mono.Unix.Native.SocketTest` tests, which started failing in
    commit ce4eeb4f with the bump to mono/mono/2019-02@c6edaa62:

    Before (passes): https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/560/testReport/(root)/SocketTest/
      After (fails): https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/561/testReport/(root)/SocketTest/

 2. Some tests which the Xamarin.iOS team flagged as a frequent cause
    of `OutOfMemoryException`-related errors.

In order to better support (2), update `XUnitTestRunner.IsIncluded()`
to support "class-based" test exclusions.  Similar to 333b98b3 and the
"Unit test exclusion" section, extend `XUnitFilterType.TypeName`
checking so that if a test name starts with an excluded name followed
by a `.`, the test is also skipped.  This allows excluding tests by
only providing the class name that contains the tests.