Stop using dlopen/dlsym to load Mono symbols
Since the dawn of time `Xamarin.Android` (nee `MonoAndroid`) used `dlopen/dlsym`
to load Mono runtime symbols that were required for the managed applications to
work on Android. This was done due to the fact that there had been no way to
reliably load shared libraries referenced by the XA runtime in earlier versions
of Android. However, thanks to the recent switch to API level 16 as our lowest
supported level, we are now able to simply preload the shared libraries we
depend on and, thus, allow the system linker to resolve all the symbols when
loading `libmonodroid.so`. This, in turn, means we can now directly link
`libmonosgen-2.0` on all target platforms (including all of the supported host
operating systems) and, what's also very important, include Mono headers
directly in our source. This makes sure we don't divert in definitions of
various macros and functions that we use accross our runtime and that we
continue to work with any version of Mono shipped in the Mono SDK archives.

This commit removes all traces of the `DylibMono` class and it also renames all
the C++ header files to have the `.hh` extension, to mirror the `.cc` source
file convention and cleary mark the C++ headers as such, as opposed to the
handful of C headers (`.h`) we have in our tree.