[Xamarin.Android.Build.Tests] Update Aapt2 tests to use the XA version.
Commit 4ccbeb74 added support for downloading `aapt2`
from maven. One thing we missed was updating the
unit tests for `aapt2` to use this rather than the
one in `build-tools`.

This commit updates those tests to use the In-Tree
version of `aapt2` rather than the one in `build-tools`.
It also fixed one of the tests as the new `aapt2` is
now raising two errors rather than just one.