[.gitattributes] overhaul line ending settings
I keep hitting an annoyance on Windows, such as:

    > git rebase master
    Cannot rebase: You have unstaged changes.
    Please commit or stash them.

What? I didn't change anything?

`git diff` reveals:

    warning: CRLF will be replaced by LF in ThirdPartyNotices.txt.
    The file will have its original line endings in your working directory

So `xaprepare` is probably just using `WriteLine()` to generate this
file. That is right and good.

The problem is our `.gitattributes` file needs to:

* Check out `.txt` files as crlf on Windows
* Commit `.txt` files as lf to the git repo

I took this opportunity to overhaul our `.gitattributes` file a bit:

* More comments / explanation
* Alphabetically sorted & used spaces for better formatting
* Configured Unix-specific files as LF