Refresh a few Android SDK components (#3213)
Update the versions of various Android SDK components:

  * Bump Android emulator package to 29.0.11.
  * Use SDK Tools 29.* instead of 28.*.
  * Use `android_m2repository` r47 instead of r16 (?!)
  * Update Android SDK package revisions used, except for API-28.

Include additional files in the `xa-build-status*.zip` and
`xa-test-results*.zip` files.

Improve Android SDK package version checking so that when a package
is updated (e.g. `pkgRevision` change), the package is updated.
(This is why `pkgRevision` is now tracked separately.)

On "interesting" SNAFU happened in the lifetime of this commit: it
originally included a bump to API-28 r6 (`platform-28_r06.zip`),
which [subsequently broke `Mono.Android.dll` compilation][0]:

	src\Mono.Android\obj\Release\android-28\mcw\Java.Lang.StringBuffer.cs(259,32): Error CS0111: Type 'StringBuffer' already defines a member called 'Append' with the same parameter types

The immediate cause of this error is that API-28 *removed* the
`package`-private type `java.lang.AbstractStringBuilder`, which broke
our binding (and is something we need to separately address):

	# From platform-28_r04.zip (currently used):
	$ jar tf ~/android-toolchain/sdk/platforms/android-28/android.jar | grep AbstractSt
	java/lang/AbstractStringBuilder.class

	# From platform-28_r06.zip:
	$ jar tf android.jar | grep AbstractStr
	# no matches

However, there were unanticipated ramifications: because previous
versions of `xaprepare` didn't know about package revisions, they
didn't "downgrade" the Android SDK installed on the machine.
Consequently, *other PRs* would fail with the same CS0111 error
because they were on a machine with API-28 r06 installed.
(Which is why it's paramount we get a version of this commit in
sooner rather than later, so that we can "un-botch" our build bots.)

Finally, [address][1] the rando [`EmbeddedDSOUnitTests` failures][2],
which are caused because the `x86_64` arch files aren't consistently
installed.  This in turn happens because `make prepare jenkins` is
unreliable, because `make jenkins` requires
`bin/Build$(Configuration)/rules.mk`, which is created by
`make prepare`.  There is thus an ordering problem: `make jenkins`
must happen *after* `make prepare` has *completed* so that it can
properly import `rules.mk`.  The fix is to execute `make jenkins`
separately from `make prepare`.

[0]: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=2767236&view=logs&jobId=cac0e8d3-0ef5-5d2b-b57e-e8fde7204df3
[1]: https://xamarinhq.slack.com/archives/C03CEGRUW/p1560518047300700
[2]: https://jenkins.mono-project.com/job/xamarin-android-pr-pipeline-release/1299/testReport/