[xaprepare] Fix some SDK component caching
A few of our Android SDK components do not specify a package revision.
This breaks caching and updating for these components. If a local version
of one of these components is already installed we won't attempt to
replace it if it is out of date. As a result, we've likely been sitting
on out of date platform-tools (adb.exe), and emulator versions on many of
our build/test machines. These caching fixes will hopefully help address
some of the emulator performance and adb.exe related crashes we've seen.