[xaprepare] Add `xaprepare -refresh` option (#3228)
Add an `xaprepare -refresh` option which provides a way to force
reinstallation of certain programs or dependencies.  Usage:

	xaprepare.exe -refresh=sdk,ndk

Usage of `xaprepare -refresh` can be enabled via **make**(1) by
setting the `$(XA_FORCE_COMPONENT_REFRESH)` value to `true`:

	make prepare XA_FORCE_COMPONENT_REFRESH=true

A few of our Android SDK components don't specify a package revision.
This breaks caching and updating for these components.  If a local
version of one of these components is already installed we won't
attempt to replace it if it is out of date.  As a result, we've
likely been sitting on out of date `platform-tools` (`adb`?), and
`emulator` versions on many of our build/test machines.

These caching fixes will hopefully help address some of the emulator
performance and `adb` related crashes we've seen.