Add a solution file for remap-assembly-ref (#3248)
We've seen the following build error when `remap-assembly-ref` is built
from within `xaprepare` by passing its project file to msbuild:

    (ValidateSolutionConfiguration target) ->
    /Users/vsts/agent/2.153.2/work/1/s/external/mono/external/cecil/Mono.Cecil.sln.metaproj :
     error MSB4126: The specified solution configuration "net_4_0_Debug|Any CPU" is invalid.
     Please specify a valid solution configuration using the Configuration and Platform properties
     (e.g. MSBuild.exe Solution.sln /p:Configuration=Debug /p:Platform="Any CPU") or leave those
     properties blank to use the default solution configuration.
     [/Users/vsts/agent/2.153.2/work/1/s/external/mono/external/cecil/Mono.Cecil.sln]

The reason is that the `$(Configuration)` property, even though set on the msbuild
command line, is reset somewhere in one of the Cecil projects to the value shown in
the above error message. This happens because properties set on command line when
building a `.csproj`, instead of an `.sln`, are **not** made read-only. Thus, using
a solution file should fix the problem and this is what this commit is adding.