Verify each Android SDK package before attempting to install (#3247)
Sometimes packages may get corrupted on disk in which case we should delete them
from the cache directory and re-download them. The verification step should be
done before installation starts, because otherwise an installation failure
exception will be thrown and the entire preparation process will be aborted.

Additionally, if unpacking of a component fails delete the cached archive as
well.