[build] Always check if we have every bundle item installed (#3253)
It may happen that some files which are supposed to be in the bundle
are missing.  In such case we need to re-download or re-install the
bundle to fix the issue.  This commit makes sure it happens.

Additionally, since all the bundle items are specified in
`Runtimes.cs`, the file must be used when calculating the bundle
"version" hash or we might face the issue when the list of expected
bundle items (in `Runtimes.cs`) doesn't agree with what's in the
currently selected bundle - because the hash didn't change even
though we added/removed bundle items in `Runtimes.cs`.