Windows build fixes
Don't include designer certain files in BuildItems on Windows

Windows, as a host OS, does NOT currently install any BCL files - it takes
everything from the bundle produced on macOS. However, the code to populate
designer BCL file sets also ran on Windows thus producing this, false negative,
error:

     bin\Debug\lib\xamarin.android\xbuild\Xamarin\Android\Windows\bcl\I18N.dll missing, skipping the rest of bundle item file scan
     Some bundle files are missing, download/rebuild/reinstall forced
     Step Xamarin.Android.Prepare.Step_PrepareBundle failed
      System.InvalidOperationException: Step Xamarin.Android.Prepare.Step_PrepareBundle failed

Fix the issue by not populating the designer BCL files on Windows.

Another issue is Windows trying to include **host** versions of the
cross-compilers in the bundle items, thus resulting in this error:

     bin\Debug\lib\xamarin.android\xbuild\Xamarin\Android\Windows\cross-arm missing, skipping the rest of bundle item file scan
      Some bundle files are missing, download/rebuild/reinstall forced

Fix the issue by ignoring all the cross-compilers on Windows.

Reset file attributes if they prevent deletion of a temporary dir

The problem, so far, occurred only on Windows, but it may happen on other
systems as well. Whenever a file is marked read-only the `Directory.Delete`
method will throw a "permission denied" exception, refusing to delete the file.
This happens, for instance, with our current bundle in which the
`libzip.5.0.dylib` file is marked read-only after 7z writes it to the system (in
my case the macOSX permissions for the file were `444` octal, so it is, indeed,
read-only).

In order to work around this we catch the `UnauthorizedAccessException` and
attempt to recursively reset permissions of all files found in the directory
being moved to `FileAttributes.Normal`.

Additionally, `MoveDirectoryContentsRecursively` now has a way to ignore all
directory deletion errors. By default the errors aren't ignored with the
exception of the Xamarin.Android bundle unpacking.