Don't include designer certain files in BuildItems on Windows
Windows, as a host, does NOT currently install any BCL files - it takes
everything from the bundle produced on macOS. However, the code to populate
designer BCL file sets also ran on Windows thus producing this, false negative,
error:

   bin\Debug\lib\xamarin.android\xbuild\Xamarin\Android\Windows\bcl\I18N.dll missing, skipping the rest of bundle item file scan
   Some bundle files are missing, download/rebuild/reinstall forced

   Step Xamarin.Android.Prepare.Step_PrepareBundle failed
    System.InvalidOperationException: Step Xamarin.Android.Prepare.Step_PrepareBundle failed

Fix the issue by not populating the designer BCL files on Windows.

Another issue is Windows trying to include **host** versions of the
cross-compilers in the bundle items, thus resulting in this error:

   bin\Debug\lib\xamarin.android\xbuild\Xamarin\Android\Windows\cross-arm missing, skipping the rest of bundle item file scan
    Some bundle files are missing, download/rebuild/reinstall forced

Fix the issue by ignoring all the cross-compilers on Windows.