[Xamarin.Android.Build.Tasks] #deletebinobj "regression protection" (#3181)
Commit 8cd3c154 fixed a #deletebinobj problem found when building a
project with Visual Studio 2017 15.9, upgrading to Visual Studio 2019
16.2 (or later), and building the project without cleaning it first.

Now it's time for the regression test for this scenario:

 1. Created the Master/Detail Xamarin.Forms template from
    Visual Studio 2017 15.9
 2. Build it in `Debug` and `Release` configurations
 3. Zip it up, and upload to Azure

I created 2 zips: Windows & Darwin.  I also added a "revision" in the
URL, for changing the zip over time.  It will get cached on a CDN, so
the URL needs to change sometimes.

Next I created an `ExistingProject` type in `Xamarin.ProjectTools`,
so I could unzip these files and build them.

To fix other weirdness with NuGet when passing `bin` and `obj` from
one machine to another, I had to:

  * Explicitly run `Restore`
  * Set `RestoreNoCache=True`, or Mac would fail to `Restore`

I plan to make further changes to `DeleteBinObjTest.cs` in other PRs.
It should start apps and connect the debugger like we are doing in
other tests.