[docs] Adjust Commercial Xamarin.Android aka.ms links
Switch all of the download.microsoft.com links to use aka.ms redirection
links, and include the branch name in each pair of aka.ms links instead
of only tracking the latest release and preview versions with aka.ms
links.  This means the links will remaining constant for any given
branch while the branch is in preview, when it moves to release, and
when it is replaced by a new release.  That in turn means that the
README won't need as many edits and that advanced users can construct a
link to a particular version based on the branch name if they want.