[docs] Add OSS Xamarin.Android master download location (#3240)
Update the Downloads documentation:

  * Remove the Xamarin.Android 7.x download links; those are ancient.
  * Add download links for Commercial Xamarin.Android 9.3 and 9.4
  * Add a download link for the OSS master build, currently 9.4.99.

Additionally, update all of the `download.microsoft.com` links to use
`aka.ms` redirection links, and include the branch name in each pair
of `aka.ms` links.  This means the links will remaining constant for
any given branch while the branch is in preview, when it moves to
release, and when it is replaced by a new release.  That in turn
means that `README.md` won't need as many edits and that advanced
users can construct a link to a particular version based on the
branch name if they want.