[Xamarin.Android.Build.Tasks] Revert special handling of keywords in Android resource names
Fixes: https://github.com/xamarin/xamarin-android/issues/3250

Partially reverts [af61ecda][0].

It turns out that the System.CodeDom APIs take care of adding `@` as
needed on keywords when outputting C# source code.  Behind the scenes at
run time, the name of the field doesn't actually have the `@` prefix
(e.g., `nameof (@fixed)` returns `"fixed"`).  So the code in af61ecda
that added an `@` prefix unintentionally led to incorrect `false` return
values from `resourceFields.Contains()` in
`ResourceDesignerImportGenerator.CreateImportFor()`.

Revert the part of af61ecda that was adding an `@` prefix.  This goes
back to letting System.CodeDom take care of adding the `@` as needed and
restores the previous correct behavior of `resourceFields.Contains()`.

Before this change, the new updated `GenerateDesignerFileFromRtxt` test
failed because the generated `UpdateIdValues()` method in the output
`Resource.designer.cs` file did not include a line for the `fixed`
library project resource.  After this change, the test passes.

In a real app, this problem would mean that the original temporary
resource ID generated in the context of the library project would be
left assigned to the C# name of the resource, so a line similar to the
following would fail:

    Resources.GetString (Library.Resource.String.@fixed);

The exception would be similar to:

    Android.Content.Res.Resources+NotFoundException: String resource ID #0x7f0a0028

[0]: https://github.com/xamarin/xamarin-android/commit/af61ecda302d69dd714f3a363829c3a9e593f5ab