[Xamarin.Android.Build.Tasks] Add a new GenerateDesignerFileFromRtxt test case with a library reference
The code path in the `<GenerateResourceDesigner/>` task that uses the
`ResourceDesignerImportGenerator` class only runs if the `@(References)`
item list includes an assembly that is not a framework assembly.

Enable testing of that code path by expanding the
`GenerateDesignerFileFromRtxt` test to include both a version that runs
without a library reference (as before) and a new version that runs
*with* a library reference.

Modify the `XamarinAndroidLibraryProject` class to set the
`$(AndroidResgenFile)` property to the current default value used in the
library project template in Visual Studio.  This property is required to
get the expected behavior from the new test.  Without this change, the
library project would build skip processing the resources (because the
`$(_AndroidResourceDesignerFile)` property would be blank).  There is a
chance that changing this property in the `XamarinAndroidLibraryProject`
class itself might break other tests.  Depending on what happens with
the other tests, it might be better to set this property on just the one
*instance* of `XamarinAndroidLibraryProject` in the new test.

Move the `menu_settings` string from the app project resources into the
library project resources so that the new library project assembly will
build successfully.  I believe this also more closely matches how a real
project would look because the library project's `menu` resource
references that string, so the string needs to be available to the
library project and not just the app project.

Add a string to the library project resources that uses a C# keyword
(`fixed`) as its name to test that using C# keywords as resource names
is handled correctly.