[monodroid] Allow loading assemblies from memory
Add the necessary code and entry-point to be able to supply assembly data directly from memory.

The motivation is to allow the designer more flexibility in how it acquire those assemblies to begin with and avoid a complicated dance of file paths juggling and file copying that are detrimental to performance on Windows and cause file locking problems.

This is achieved by introducing a new class `InMemoryAssemblies` that, much like `EmbeddedAssemblies`, can be queried when a new .dll needs to be loaded to fetch it from a store that's keyed per-assembly name and per-appdomain. If not found in this new store it falls back to the existing logic that can load assemblies directly from disk.

Since we only care about this capability on desktop, the support is only compiled for host platforms (both Mac and Windows).