[Xamarin.Android.Build.Tasks] Fix up DesignTime build Resource.Designer.cs
One of the problems we have with the `ManagedResourceParser` is
that it uses the `R.txt` to generate the `Resource.Designer.cs`
file. The issue there is that if the user modifies a layout file
(say adds or changes and id) that will NOT be picked up until
the next build. This is because the `R.txt` file is generated
by `aapt/aapt2` and those tools are not called for DesignTime
Builds.

So what we need to do is pass in the `$(_AndroidResFlagFile)`
into the `GenerateResourceDesigner`. This will allow us to
do a last modified date time check against the `R.txt` file.
If the `$(_AndroidResFlagFile)` is newer then we should NOT
use the `R.txt` and parse the layout files to pick up changes.