[Xamarin.Android.Build.Tasks] Allow C# keywords as Resource names (#3275)
Fixes: https://github.com/xamarin/xamarin-android/issues/3250

Partially reverts commit af61ecda.

It turns out that the System.CodeDom APIs take care of adding `@` as
needed on keywords when outputting C# source code.  Behind the scenes
at runtime, the name of the field doesn't actually have the `@`
prefix (e.g., `nameof (@fixed)` returns `"fixed"`).  So the code in
af61ecda that added an `@` prefix unintentionally led to incorrect
`false` return values from `resourceFields.Contains()` in
`ResourceDesignerImportGenerator.CreateImportFor()`.

Revert the part of af61ecda that was adding an `@` prefix.  This goes
back to letting System.CodeDom take care of adding the `@` as needed
and restores the old correct behavior of `resourceFields.Contains()`.

Before this change, the updated
`ManagedResourceParserTests.GenerateDesignerFileFromRtxt()` test
failed because the generated `UpdateIdValues()` method in the output
`Resource.designer.cs` file did not include a line for the `fixed`
library project resource.  After this change, the test passes.

In a real app, this problem would mean that the original temporary
resource ID generated in the context of the library project would be
left assigned to the C# name of the resource, so a line similar to
the following would fail:

	Resources.GetString (Library.Resource.String.@fixed);

with an exception similar to:

	Android.Content.Res.Resources+NotFoundException: String resource ID #0x7f0a0028

Updated the `ManagedResourceParserTests.GenerateDesignerFileFromRtxt()`
test case with a library reference to cover this scenario.

The code path in the `<GenerateResourceDesigner/>` task that uses the
`ResourceDesignerImportGenerator` class only runs if the
`@(References)` item list includes an assembly that is not a
framework assembly.

Enable testing of that code path by expanding the
`GenerateDesignerFileFromRtxt()` test to include both a version that
runs without a library reference (as before) and a new version that
runs *with* a library reference.

Modify the `XamarinAndroidLibraryProject` class to set the
`$(AndroidResgenFile)` property to the current default value used in
the library project template in Visual Studio.  This property is
required to get the expected behavior from the new test.  Without this
change, the library project would build skip processing the resources,
as the `$(_AndroidResourceDesignerFile)` property would be blank.

Move the `menu_settings` string from the app project resources into
the library project resources so that the new library project assembly
will build successfully.  I believe this also more closely matches how
a real project would look because the library project's `menu`
resource references that string, so the string needs to be available
to the library project and not just the app project.

Add a string to the library project resources that uses a C# keyword
(`fixed`) as its name to test that using C# keywords as resource names
is handled correctly.